requirements.txt
pytest
pyautogui
pytest-html-reporter
marshmallow
pytest-bdd
pyscreenrec
pytest-yaml
pytest-testrail
pytest-rerunfailures
pytest-html
requests
pytest-dependency
pytest-html-reporter